requirements.txt
binance
pandas
numpy
matplotlib
python-binance
requests
ta
scipy
scikit-learn
seaborn
jupyter
sqlalchemy
TA-Lib
Pillow
mplfinance
schedule
